requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
sqlparse==0.5.3
django-filter==5.2.7
dj-database-url==3.0.1